+++ frontend/byte-size/app/home/page.tsx
@@ -1,5 +1,4 @@
 import Card from "@/components/home/card";
-import { DEPLOY_URL } from "@/lib/constants";
 import { Github, Twitter } from "@/components/shared/icons";
 import WebVitals from "@/components/home/web-vitals";
 import ComponentGrid from "@/components/home/component-grid";
@@ -39,6 +38,7 @@ export default async function UserHome() {
           <Card
             key={title}
             title={title}
+            demo={null}
             // description={description}
             repoTitle={title}
             large={large}
@@ -62,52 +62,4 @@ const features = [
     title: "LLaMa2",
     large: true,
   }
-  // {
-  //   title: "Performance first",
-  //   description:
-  //     "Built on [Next.js](https://nextjs.org/) primitives like `@next/font` and `next/image` for stellar performance.",
-  //   demo: <WebVitals />,
-  // },
-  // {
-  //   title: "One-click Deploy",
-  //   description:
-  //     "Jumpstart your next project by deploying Precedent to [Vercel](https://vercel.com/) in one click.",
-  //   demo: (
-  //     <a href={DEPLOY_URL}>
-  //       <Image
-  //         src="https://vercel.com/button"
-  //         alt="Deploy with Vercel"
-  //         width={120}
-  //         height={30}
-  //         unoptimized
-  //       />
-  //     </a>
-  //   ),
-  // },
-  // {
-  //   title: "Built-in Auth + Database",
-  //   description:
-  //     "Precedent comes with authentication and database via [Auth.js](https://authjs.dev/) + [Prisma](https://prisma.io/)",
-  //   demo: (
-  //     <div className="flex items-center justify-center space-x-20">
-  //       <Image alt="Auth.js logo" src="/authjs.webp" width={50} height={50} />
-  //       <Image alt="Prisma logo" src="/prisma.svg" width={50} height={50} />
-  //     </div>
-  //   ),
-  // },
-  // {
-  //   title: "Hooks, utilities, and more",
-  //   description:
-  //     "Precedent offers a collection of hooks, utilities, and `@vercel/og`",
-  //   demo: (
-  //     <div className="grid grid-flow-col grid-rows-3 gap-10 p-10">
-  //       <span className="font-mono font-semibold">useIntersectionObserver</span>
-  //       <span className="font-mono font-semibold">useLocalStorage</span>
-  //       <span className="font-mono font-semibold">useScroll</span>
-  //       <span className="font-mono font-semibold">nFormatter</span>
-  //       <span className="font-mono font-semibold">capitalize</span>
-  //       <span className="font-mono font-semibold">truncate</span>
-  //     </div>
-  //   ),
-  // },
 ];
