+++ frontend/byte-size/package.json
@@ -3,21 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "prisma generate && next dev",
-    "build": "prisma generate && prisma db push && next build",
+    "dev": "next dev",
+    "build": "next build",
     "format:write": "prettier --write \"**/*.{css,js,json,jsx,ts,tsx}\"",
     "format": "prettier \"**/*.{css,js,json,jsx,ts,tsx}\"",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
-    "@next-auth/prisma-adapter": "^1.0.7",
-    "@prisma/client": "^5.6.0",
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-popover": "^1.0.7",
     "@radix-ui/react-tooltip": "^1.0.7",
-    "@types/node": "18.11.18",
-    "@types/react": "18.2.24",
     "@types/react-dom": "18.2.8",
     "@vercel/analytics": "^0.1.11",
     "@vercel/og": "^0.0.26",
@@ -35,22 +31,22 @@
     "react-dom": "18.2.0",
     "react-markdown": "^8.0.7",
     "tailwind-merge": "^1.14.0",
-    "typescript": "5.2.2",
     "use-debounce": "^9.0.4",
-    "vaul": "^0.6.8",
-    "vercel": "^32.6.0"
+    "vaul": "^0.6.8"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.6",
     "@tailwindcss/line-clamp": "^0.4.4",
     "@tailwindcss/typography": "^0.5.10",
     "@types/ms": "^0.7.32",
+    "@types/node": "18.11.18",
+    "@types/react": "18.2.24",
     "autoprefixer": "^10.4.16",
     "concurrently": "^7.6.0",
     "postcss": "^8.4.31",
     "prettier": "^2.8.8",
     "prettier-plugin-tailwindcss": "^0.2.8",
-    "prisma": "^5.6.0",
-    "tailwindcss": "^3.3.3"
+    "tailwindcss": "^3.3.3",
+    "typescript": "5.2.2"
   }
 }
