+++ frontend/byte-size/tsconfig.json
@@ -47,7 +47,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "app/api/auth/[...nextauth]/route.ts"
+    "app/api/auth/[...nextauth]/route"
   ],
   "exclude": [
     "node_modules"
